Disable FS event notifier test on mac and consolidate fs tests.
 - FSEvent order isn't predictable on mac causing frequent failures
 - moved test to the core module and removed duplicates since the test
   itself is platform independent
 - run (now agregated) masterfs tests on all platforms (previously
   only linux) and:
   - disable the NotifierTest on mac
   - disable all other failing tests on windows (since they didn't run
     for a long time, see TODOs)
   - fix some JDK sensitive tests
